requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
docstring_parser==0.16
faiss-cpu==1.8.0.post1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-aiplatform==1.64.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.66.0
idna==3.8
numpy==1.26.4
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
